# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Textile's HTTP REST API Documentation",
        "title": "Textile REST API",
        "termsOfService": "https://github.com/b582q9/go-textile-sapien/blob/master/TERMS",
        "contact": {
            "name": "Textile",
            "url": "https://textile.io/",
            "email": "contact@textile.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/b582q9/go-textile-sapien/blob/master/LICENSE"
        },
        "version": "0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/ipfs/cat/{path}": {
            "get": {
                "description": "Displays the data behind an IPFS CID (hash) or Path",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "Cat IPFS data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ipfs/ipns cid",
                        "name": "path",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "key=",
                        "description": "key: Key to decrypt data on-the-fly",
                        "name": "X-Textile-Opts",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ipfs/id": {
            "get": {
                "description": "Displays underlying IPFS peer ID",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "Get IPFS peer ID",
                "responses": {
                    "200": {
                        "description": "peer id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ipfs/pubsub/pub/{topic}": {
            "post": {
                "description": "Publishes a message to a given pubsub topic",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "Publish a message",
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ipfs/pubsub/sub/{topic}": {
            "get": {
                "description": "Subscribes to messages on a given topic",
                "produces": [
                    "text/event-stream with events",
                    " or just application/json"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "Subscribe messages",
                "responses": {
                    "200": {
                        "description": "results stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ipfs/swarm/connect": {
            "post": {
                "description": "Opens a new direct connection to a peer using an IPFS multiaddr",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "Opens a new direct connection to a peer address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "peer address",
                        "name": "X-Textile-Args",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ipfs/swarm/peers": {
            "get": {
                "description": "Lists the set of peers this node is connected to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipfs"
                ],
                "summary": "List swarm peers",
                "parameters": [
                    {
                        "type": "string",
                        "default": "verbose=\"false\",latency=\"false\",streams=\"false\",direction=\"false\"",
                        "description": "verbose: Display all extra information, latency: Also list information about latency to each peer, streams: Also list information about open streams for each peer, direction: Also list information about the direction of connection",
                        "name": "X-Textile-Opts",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "connection",
                        "schema": {
                            "$ref": "#/definitions/ipfs.ConnInfos"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ipfs.ConnInfos": {
            "type": "object",
            "properties": {
                "peers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ipfs.connInfo"
                    }
                }
            }
        },
        "ipfs.connInfo": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "muxer": {
                    "type": "string"
                },
                "peer": {
                    "type": "string"
                },
                "streams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ipfs.streamInfo"
                    }
                }
            }
        },
        "ipfs.streamInfo": {
            "type": "object",
            "properties": {
                "protocol": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}